Development

+++ build.sh
@@ -1,10 +1,18 @@
 #!/bin/bash
 
-export FLASK_APP=run.py
-export FLASK_DEBUG=1
-rm -r -f migrations
-rm *.db
-flask db init
-flask db migrate -m "Initial migration"
-flask db upgrade
-flask run
+# Clean mode
+if [ "$1" == "--clean" ]; then
+  export FLASK_APP=run.py
+  export FLASK_DEBUG=1
+  rm -r -f migrations
+  rm *.db
+  flask db init
+  flask db migrate -m "Initial migration"
+  flask db upgrade
+else
+  # Normal mode
+  export FLASK_APP=run.py
+  export FLASK_DEBUG=1
+fi
+
+flask run --port=5004
